+++ pkg/consolegraphql/resolvers/resolver_link.go
@@ -19,7 +19,7 @@ type linkSpecK8sResolver struct{ *Resolver }
 
 func (l linkSpecK8sResolver) Connection(ctx context.Context, obj *consolegraphql.LinkSpec) (*consolegraphql.Connection, error) {
 	if obj != nil {
-		linkrsctx := graphql.GetResolverContext(ctx).Parent.Parent
+		linkrsctx := graphql.GetFieldContext(ctx).Parent.Parent
 		link := linkrsctx.Result.(**consolegraphql.Link)
 
 		namespace := (*link).ObjectMeta.Namespace